feat: implement OOP with abstract class
