+++ src/clients/UBAClient/UBAClientWithRefresh.ts
@@ -4,9 +4,9 @@ import { BigNumber } from "ethers";
 import { DepositWithBlock, FillWithBlock, RefundRequestWithBlock, UbaFlow } from "../../interfaces";
 import { HubPoolClient, SpokePoolClient } from "..";
 import { isDefined, sortEventsAscending } from "../../utils";
-import { RequestValidReturnType, UBAClientManual } from "./UBAClientWithManual";
+import { BaseUBAClient, RequestValidReturnType } from "./UBAClientAbstract";
 
-export class UBAClientWithRefresh extends UBAClientManual {
+export class UBAClientWithRefresh extends BaseUBAClient {
   // @dev chainIdIndices supports indexing members of root bundle proposals submitted to the HubPool.
   //      It must include the complete set of chain IDs ever supported by the HubPool.
   // @dev SpokePoolClients may be a subset of the SpokePools that have been deployed.
@@ -16,7 +16,7 @@ export class UBAClientWithRefresh extends UBAClientManual {
     private readonly spokePoolClients: { [chainId: number]: SpokePoolClient },
     logger?: winston.Logger
   ) {
-    super(chainIdIndices, { spoke: {} }, logger);
+    super(chainIdIndices, logger);
     assert(chainIdIndices.length > 0, "No chainIds provided");
     assert(Object.values(spokePoolClients).length > 0, "No SpokePools provided");
   }
@@ -56,21 +56,6 @@ export class UBAClientWithRefresh extends UBAClientManual {
     return { blockNumber, spokePoolBalance };
   }
 
-  /**
-   * @description Construct the ordered sequence of SpokePool flows between two blocks.
-   * @note Assumptions:
-   * @note Deposits, Fills and RefundRequests have been pre-verified by the SpokePool contract or SpokePoolClient, i.e.:
-   * @note - Deposit events contain valid information.
-   * @note - Fill events correspond to valid deposits.
-   * @note - RefundRequest events correspond to valid fills.
-   * @note In order to provide up-to-date prices, UBA functionality may want to follow close to "latest" and so may still
-   * @note be exposed to finality risk. Additional verification that can only be performed within the UBA context:
-   * @note - Only the first instance of a partial fill for a deposit is accepted. The total deposit amount is taken, and
-   * @note   subsequent partial, complete or slow fills are disregarded.
-   * @param spokePoolClient SpokePoolClient instance for this chain.
-   * @param fromBlock       Optional lower bound of the search range. Defaults to the SpokePool deployment block.
-   * @param toBlock         Optional upper bound of the search range. Defaults to the latest queried block.
-   */
   public getFlows(chainId: number, fromBlock?: number, toBlock?: number): UbaFlow[] {
     const spokePoolClient = this.spokePoolClients[chainId];
 
@@ -119,15 +104,6 @@ export class UBAClientWithRefresh extends UBAClientManual {
     return flows;
   }
 
-  /**
-   * @description Evaluate an RefundRequest object for validity.
-   * @dev  Callers should evaluate 'valid' before 'reason' in the return object.
-   * @dev  The following RefundRequest attributes are not evaluated for validity and should be checked separately:
-   * @dev  - previousIdenticalRequests
-   * @dev  - Age of blockNumber (i.e. according to SpokePool finality)
-   * @param chainId       ChainId of SpokePool where refundRequest originated.
-   * @param refundRequest RefundRequest object to be evaluated for validity.
-   */
   public refundRequestIsValid(chainId: number, refundRequest: RefundRequestWithBlock): RequestValidReturnType {
     const { relayer, amount, refundToken, depositId, originChainId, destinationChainId, realizedLpFeePct, fillBlock } =
       refundRequest;
